Remove unnecessary type cast
Commit UUID: cf91f2ee-bad0-471c-a245-842aa121182a

+++ trunk/schema_parse.c
@@ -214,7 +214,7 @@ get_node_id(const struct radix_tree_node *node)
         init = 1;
     }
 
-    return xtea_encrypt((uint64_t)(uintptr_t)node, key);
+    return xtea_encrypt((uintptr_t)node, key);
 }
 
 static int
